+++ src/main/kotlin/uoslife/servermeeting/domain/meeting/domain/service/impl/SingleMeetingService.kt
@@ -7,24 +7,20 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import uoslife.servermeeting.domain.meeting.application.response.MeetingTeamInformationGetResponse
 import uoslife.servermeeting.domain.meeting.application.response.MeetingTeamUserListGetResponse
-import uoslife.servermeeting.domain.meeting.application.response.UserProfile
 import uoslife.servermeeting.domain.meeting.domain.dao.InformationUpdateDao
 import uoslife.servermeeting.domain.meeting.domain.dao.PreferenceUpdateDao
 import uoslife.servermeeting.domain.meeting.domain.dao.UserTeamDao
-import uoslife.servermeeting.domain.meeting.domain.entity.Information
 import uoslife.servermeeting.domain.meeting.domain.entity.MeetingTeam
-import uoslife.servermeeting.domain.meeting.domain.entity.Preference
 import uoslife.servermeeting.domain.meeting.domain.entity.enums.TeamType
 import uoslife.servermeeting.domain.meeting.domain.exception.*
 import uoslife.servermeeting.domain.meeting.domain.repository.InformationRepository
 import uoslife.servermeeting.domain.meeting.domain.repository.MeetingTeamRepository
 import uoslife.servermeeting.domain.meeting.domain.repository.PreferenceRepository
 import uoslife.servermeeting.domain.meeting.domain.service.BaseMeetingService
+import uoslife.servermeeting.domain.meeting.domain.service.util.MeetingServiceUtils
 import uoslife.servermeeting.domain.meeting.domain.util.Validator
-import uoslife.servermeeting.domain.user.domain.entity.User
 import uoslife.servermeeting.domain.user.domain.exception.UserNotFoundException
 import uoslife.servermeeting.domain.user.domain.repository.UserRepository
-import java.time.LocalDate
 import java.util.*
 
 @Service
@@ -38,6 +34,7 @@ class SingleMeetingService(
     private val userTeamDao: UserTeamDao,
     private val preferenceUpdateDao: PreferenceUpdateDao,
     private val informationUpdateDao: InformationUpdateDao,
+    private val meetingServiceUtils: MeetingServiceUtils,
     private val validator: Validator,
     @Value("\${app.season}")
     private val season: Int,
@@ -85,8 +82,9 @@ class SingleMeetingService(
         val information = informationRepository.findByMeetingTeam(meetingTeam)
         val preference = preferenceRepository.findByMeetingTeam(meetingTeam)
 
-        informationUpSert(information, meetingTeam, informationDistance, informationFilter, informationMeetingTime)
-        preferenceUpSert(preference, meetingTeam, preferenceDistance, preferenceFilter)
+        meetingServiceUtils
+            .informationUpSert(information, meetingTeam, informationDistance, informationFilter, informationMeetingTime)
+        meetingServiceUtils.preferenceUpSert(preference, meetingTeam, preferenceDistance, preferenceFilter)
     }
 
     override fun getMeetingTeamInformation(userUUID: UUID): MeetingTeamInformationGetResponse {
@@ -98,7 +96,10 @@ class SingleMeetingService(
         val information = informationRepository.findByMeetingTeam(meetingTeam) ?: throw InformationNotFoundException()
         val preference = preferenceRepository.findByMeetingTeam(meetingTeam) ?: throw PreferenceNotFoundException()
 
-        return toMeetingTeamInformationGetResponse(user, information, preference)
+        return meetingServiceUtils
+            .toMeetingTeamInformationGetResponse(
+                user.gender, TeamType.SINGLE, listOf(user), information, preference, null
+            )
     }
 
     @Transactional
@@ -109,89 +110,5 @@ class SingleMeetingService(
         val meetingTeam = userTeam.team
 
         meetingTeamRepository.deleteById(meetingTeam.id!!)
-    }
-
-    fun toMeetingTeamInformationGetResponse(
-        user: User,
-        information: Information,
-        preference: Preference,
-    ): MeetingTeamInformationGetResponse {
-        val currentYear: Int = LocalDate.now().year
-        val userBirthYear: Int = user.birthYear as Int
-        return MeetingTeamInformationGetResponse(
-            sex = user.gender,
-            teamType = TeamType.SINGLE,
-            teamName = null,
-            informationDistance = information.distanceInfo,
-            informationFilter = information.filterInfo,
-            informationMeetingTime = information.meetingTime,
-            preferenceDistance = preference.distanceCondition,
-            preferenceFilter = preference.filterCondition,
-            teamUserList = listOf(
-                UserProfile(
-                    nickname = user.nickname,
-                    age = currentYear - userBirthYear + 1,
-                    kakaoTalkId = user.kakaoTalkId!!,
-                    department = user.department!!,
-                    studentType = user.studentType!!,
-                    height = user.height,
-                    smoking = user.smoking,
-                    spiritAnimal = user.spiritAnimal,
-                    MBTI = user.mbti!!,
-                    interest = user.interest,
-                ),
-            ),
-        )
-    }
-
-    // information 이 없으면 생성, 있으면 수정
-    private fun informationUpSert(
-        information: Information?,
-        meetingTeam: MeetingTeam,
-        informationDistance: String,
-        informationFilter: String,
-        informationMeetingTime: String,
-    ) {
-        if (information == null) {
-            informationRepository.save(
-                Information(
-                    meetingTeam = meetingTeam,
-                    filterInfo = informationFilter,
-                    distanceInfo = informationDistance,
-                    meetingTime = informationMeetingTime,
-                ),
-            )
-        } else {
-            informationUpdateDao.updateInformationByMeetingTeam(
-                meetingTeam,
-                informationDistance,
-                informationFilter,
-                informationMeetingTime,
-            )
-        }
-    }
-
-    // preference 가 없으면 생성, 있으면 수정
-    private fun preferenceUpSert(
-        preference: Preference?,
-        meetingTeam: MeetingTeam,
-        preferenceDistance: String,
-        preferenceFilter: String,
-    ) {
-        if (preference == null) {
-            preferenceRepository.save(
-                Preference(
-                    meetingTeam = meetingTeam,
-                    filterCondition = preferenceFilter,
-                    distanceCondition = preferenceDistance,
-                ),
-            )
-        } else {
-            preferenceUpdateDao.updatePreferenceByMeetingTeam(
-                meetingTeam,
-                preferenceFilter,
-                preferenceDistance,
-            )
-        }
     }
 }
